Add jenkins-ci.org YUM repo GPG Key
Without this patch the jenkins module has basic support for CentOS 5 but
cannot verify GPG signatures of the RPM packages.

This patch adds the GPG public key directly into the module, manages the
key on disk as a file resource, and imports the key into the RPM
database using an exec resource.