Add jenkins::git inert class by default
This change changes the jenkins module to define a new jenkins::git
class.  The new jenkins::git class uses the already defined
install-jenkins-plugin defined resource type to install the Git plugin.
The change should automatically restart the jenkins service when Puppet
manages the new plugin.

The change itself is inert to preserve the existing behavior of the
jenkins module.  This should not be an API breaking change.  An example
of how to take advantage of this new class is provided in the
tests/site.pp file.