(maint) Add Modulefile for puppet-module tool and forge
Without this patch, no module metadata file exists.  This makes using
the module with the puppet-module [1] tool difficult.  This patch adds a
basic Modulefile to work with the puppet-module build command.

The Modulefile needs to be updated with the LICENSE information for this
module.  I could not figure it out from the repository or commit
history.

With this patch, this module may be built using:

    % rake build
    ===========================================================
    Building /Users/jeff/vms/puppet/modules/jenkins for release
    -----------------------------------------------------------
    Done. Built: pkg/rtyler-jenkins-0.0.1.tar.gz

The built module may be installed using the puppet module tool.  The
built module package may also be uploaded to the Forge.