Add rspec tests, README, and Rakefile
Without this patch no behavior tests exist for the Jenkins module.  This
patch lays the foundation for spec testing using the rspec-puppet gem.
A Rakefile with a default task of spec provides the command line
interface to running the spec test suite.

This patch requires the rspec-puppet 0.1.0 gem to be installed and
available to Ruby.  Only one test exists to exercise the testing
framework.  Rspec will validate a Puppet manifest with a simple "include
jenkins" statement will compile into a catalog which contains the
jenkins class.

I've tested the spec testing with the following integration:

 * puppet v2.7.6
 * facter v1.6.3
 * stdlib v2.2.0
 * rspec-puppet v0.1.0

The README is updated by this patch to describe how to run the test
suite and install necessary dependencies.