Fix jenkins::firewall in case you don't have puppetlabs-firewall.
In case you don't have puppetlabs-firewall, the defined('firewall') still evaluates to true because of scoping: there's a firewall class in the jenkins namespace (indeed, we're inside that jenkins::firewall class). Puppet therefore complains that the type Firewall doesn't exist on the next line.
Using defined('::firewall') fixes the issue, and for consistency let's use ::firewall on the next line too.