Fix Could not find group jenkins error
Without this patch applied, the Jenkins module fails on Lucid when
managing a jenkins::plugin::install resource.  This is a problem because
the catalog cannot compile and the system is not configured.

The error looks like:

    err: /Stage[main]/Site::Jenkins_plugins/Jenkins::Plugin::Install[git]/File[/var/lib/jenkins]:
        Could not evaluate: Could not find group jenkins

This error is caused because the file resource specfies group =>
jenkins, but the defined type does not manage a Group[jenkins] resource,
only a User[jenkins] resource.

This patch fixes the problem by managing Group[jenkins] similar to how
User[jenkins] is already managed.

This fixes the issue and adds rspec-puppet tests to cover the common
resource of Jenkins::Plugin::Install[git]