Respect ui_{user,pass} in upstart template
When the sysv script was converted to upstart in 149f22d it forgot about the
handling of `ui_user_flag` and `ui_pass_flag` variables. Thus the `ui_user`
and `ui_pass` params had no effect.

Added a simple spec test for this particular issue. It should be extended
for the other params/variables used by the template.