Don't manage API token for jenkins::api_user
So it turns out that the API token is hashed against both the username
(which is fine) and a seed that is specific to that installation of Jenkins
(not so fine). Meaning that the hash of an API token will invalidated
whenever a Jenkins master node is rebuilt. As such, it makes no sense for us
to set this. We should just create the user and then fetch the current token
from the Jenkins UI.

The file is now owned by `jenkins` and not replaced if it already exists.