Optionally manage plugin configuration with puppet
The way this code is intended to be used is when managing a plugin and plugin configuration for a production Jenkins instance, using the files generated in a web UI of a staging instance.

For example, if you have a test/staging instance of Jenkins where you validate
your plugin configuration, you could take the myplugin.xml conifguration file
from that machine, add it to your Git repository and reference it here.