confine fact to only run on linux boxes
Confining this fact required moving the conditional and variables inside the `setcode` block, however the overall logic of the fact is unchanged. This change also includes the recent patch 7c03e35ce4f5d2c7a112142cf58cb714f95dff02

I've tested this on Linux Jenkins box, and on a Windows box and it's working as expected.