Use a more explicit scope in the custom fact when referring to 'self'
Because of the way facter pulls in code blocks (Procs) inside of Facter#add we
cannot properly refer to `self` from within the block.

In the process of fixing this, noticed one other issue with the reworked Fact,
corrected that while I was at it.

Tested with `RUBYLIB=$PWD/lib FACTERLIB=$PWD/lib facter -d 2>&1 | grep jenkins`

Fixes #126