Add defined check for plugin_parent_dir resource
The management of the basic dependencies of a Jenkins plugin should not
be unnecessarily forced on the user. This change wraps the plugin_parent_dir
resource creation in a `defined` function to allow the user to create the
Jenkins home directory themselves if necessary.

This fixes #148 in a more backwards compatible way than #149.