Add support for "No Proxy Host" list
The Jenkins proxy configuration supports a whitelist of nodes that will
not use the the proxy when Jenkins (or some plugins) attempts to
communicate. As of right now, this setting will be cleared by Puppet
if any other aspect of the proxy configuration is managed.

This commit adds support for the "No Proxy Whitelist" by adding a new
parameter to the `jenkins` class. The new parameter, `no_proxy_list`,
will accept an array of hostname patterns to no longer use the proxy to
access. The `no_proxy_list` parameter is optional and will default to
`undef`.