Improve how `::jenkins::proxy` accesses variables
Without this commit, the template referenced by the class
`::jenkins::proxy` was using the `lookupvar()` method in the ERB
tags to access variables from other class scopes. This method of
accessing the variables is at the mercy of changes to Puppet's internal
Ruby API. This method is also not very clear as to what is happening
without having to read through both the class and the template.

This commit changes the method of variable access to use a simple
variable assignment operation in the Puppet DSL to bring the values of
the required variables into the local scope. This is preferred as it
makes the template easier to read and keeps the template's operations
relative to the place from where it is called. In the Puppet code
itself, it is now more clear as to what data is used by the
`::jenkins::proxy` class and where that data is coming from.