Make creation of user optional
Currently, the creation of jenkins user
and group is protected by if !defined. This
can be extremely annoying in cases where code
is trying to install a competing jenkins user/group.

In my specific case, I'm blocked by this code
leading to duplicate jenkins users (and I need
my definition to be created first). In general,
I thought that these kinds of issue were resolvable
via enusre that classes were including in a certain
order, but I cannot seem to resove it in my specific use
case which involves:
* nested class includes
* serialized user resources from hiera
* create_resources