Cleanup metadata.json for better mechanical score
Puppet Forge has been marking the mechanical metadate score down because
of issues with the metadata.json

This change cleans up the project page URL so that it is verifiable and
will actually work correctly when clicking on the link from Puppet Forge

This change modifies the license clause to be SPDX compliant

This change adds bounds to the dependencies as having unbounded ranges
lowers the score. The current bounds should be large enough to handle
future versions properly for at least a little while.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>