Upgraded puppetlabs-apt version
PuppetLabs launched the 2.1.0 branch last month. Although it is a
"major" upgrade, none of the major changes are relevant to this module
(which only takes advantage of the apt::source object type).

This update allows this module to run with either the old and new
version of the apt module, rather than forcing a specific version.