adds support for apt module v1 and v2
Users of this module should select which version of the apt-module
they're using with the "apt_version" parameter.

This parameter is set to "1" for backwards compatability- users who
update this module should not see any issues. However, in the future
(presumably 2.0 if semver is being used) it may make sense to switch
this to 2.