enable install of cli_helper when jenkins::cli => true
+ make true the default value of the `jenkins` class `cli` param

This modules uses the cli jar + groovy helper for many of its'
management functions and is currently relying on these to be
automatically included by dependent resources.  For most use cases of
this module, the cli jar + groovy helper are automatically installed
regardless of the value of the `cli` param.  However, this does not
work for the experimental native types which are dependent on the cli
jar and groovy helper.