cleanup existing plugin archive if extension changes
At least as of jenkins 1.651, if the version of a plugin being
downloaded has a .hpi extension, and there is an existing version of the
plugin present with a .jpi extension, jenkins will actually delete the
.hpi version when restarted.  Essentially making it impossible to
(up|down)grade a plugin from .jpi -> .hpi via puppet across extension
changes.  Regardless, we should be relying on jenkins to guess which
plugin archive to use and cleanup any conflicting extensions.