pin all plugins by default
Jenkins' core plugins have an invisible "pinning" to the version that
Jenkins shipped with.  A new version must be "pinned" or jenkins will
revert to the bundled version upon restart.  This is surprising and
unexpected behavior that is very likely to confuse the end user.  To
avoid this, we should pin plugins by default.

See: https://wiki.jenkins-ci.org/display/JENKINS/Pinned+Plugins

resolves #465