attempt to determine the correct gem provider
`jenkins::cli::config` currently installs the `retries` gem needed by
the xtypes as a convenience -- it is not clear if functionality should be
in this class at all or if it should merely be a vehicle for passing
data into the catalog.  Puppetserver users will still have to manually
install the `retries` gem as there is no reliable way to determine if an
agent run is configuring a puppet master.

resolves #434
resolves #498