Added LDAP support.
By setting the user_realm to ldap and passing in the necessary config fields through hiera for example puppet will manage the ldap config.

Parameters to be set in the hash:

    server                     => "",
    rootDN                     => "",
    userSearch                 => "",
    managerDN                  => "",
    managerPasswordSecret      => "",
    inhibitInferRootDN         => "",
    disableMailAddressResolver => "",
    displayNameAttributeName   => "",
    mailAddressAttributeName   => ""

These will be loaded into a json file and then used by the puppet_helper groovy script to pass these arguments to the LDAPSecurityRealm constructor function.