Add support for disabling management of Service[jenkins]
This is especially useful when deploying Jenkins from a Docker container, but
still wishing to use some of the great abstractions around job/user/etc
management in this module.

This doesn't change the expectation that /something/ will implement
Service[jenkins] in the catalog, as that would make the implementation of types
like jenkins::cli::exec obscenely complex (and maybe not even possible).