stop jenkins-slave service before systemd unit install
If the jenkins-slave service is already running under a sysv init script
and we are converting to systemd, the sysv instance will end up being
unmanaged and likely running in parallel with the new systemd managed
instance. To prevent this, we force the jenkins-slave service to be in
the `stopped` state before even installing the systemd service unit.