enable systemd service on RedHat
Enable systemd service unit setup on $::osfamily == RedHat via the use
of jenkins::systemd.  For the time being, we are limited jenkins master
systemd support to RedHat as there is no variable interpolation in the
default sysconfig file.  Where as the Debian packaging contains
interpolation in the default /etc/default/jenkins file.  It would be
best to avoid the shenanigans that are used for swarm clients to work by
overwriting the sysconf file and/or fixing the upstream packaging.