Make all archive options accessible for plugin type
At our organization, we are downloading plugins from our own caching
mirror, but we can't access our internal mirror through our proxy
server, and we do need to authenticate to this server.

With this change, archive (and in turn, curl) will pick up the proxy
configuration from environment variables or ~/.curlrc, including the
no_proxy setting we need. When users desire to override those, they can
still specify proxy_server directly.

For authenticating with the download server, we can pass username and
password to archive.

In our private wrapper module, we set default for the jenkins::plugin
type:
```
Jenkins::Plugin {
  archive_options => |
    username => $artifactory_username,
    password => $artifactory_password,
}
```

This works because we define the defaults and the actual resources
inside the same class. Setting defaults for types is guaranteed by
Puppet only for resources declared in the same class, so we cannot
set defaults on the archive type.

If full backwards compatibility is desired, the parameter declaration
could be changed to:
```
  Hash $archive_options = { 'proxy_server' => $::jenkins::proxy::url },
```