Remove retries gem
This removes the retries gem in favor of the built in RetryAction. This
does swap the parameter cli_try_sleep for the built in exponential
backoff.

It also raises a RetriesExceeded exception instead of the
original exception. Since those aren't caught explicitly, that's not
really a huge difference. RetryAction logs at the info level which
exception has been caught.