Drop sysvinit support in jenkins::slave
For the OS families where a template was provided now default to using
systemd and we don't test with sysvinit.

This drops the ability to smoothly transition from sysvinit to systemd,
but that's not really important.  Very few users will switch their
system from sysvinit to systemd and if they do, it's a bigger upgrade
anyway.

It doesn't implement any cleanup since it's not expected to be present.