Move systemd overrides to user_setup for Debian
On Debian services are started automatically during package
installation. The setup wizard is configured via the systemd override so
creating this prior to package installation works around it. On
non-Debian (where it isn't autostarted) it doesn't matter and managing
all platforms the same has benefits.

The notify_service was missing which caused the process to run with an
incorrect parameter. This also adds a test to verify the argument is
really there when running. It should be noted that on Ubuntu 20.04 I
noticed a weird bug where it wasn't picked up. The test did catch it so
when the OS version support is added it'll be easier to catch.

Fixes: 98f46ca999635d0a9a2f4ad84d70ed60389f7b47