Update firewall.pp to replace action with jump
https://forge.puppet.com/modules/puppetlabs/firewall/readme#migration-path-to-v700

The action attribute within the firewall type has been removed as it was merely a restricted version of the jump attribute, both of them managing the same function, this being reasoned as a way to enforce the use of generic parameters. From this point the parameters formerly unique to action should now be passed to jump.