do not verify the APT key fingerprint
Verifying the APT key fingerprint should really be a responsibility of
the `puppetlabs-apt` module, so we deliberately choose to not do it
here.

Note that, currenlty, neither the (obsolete) `apt::key` nor the
`apt::keyring` resources do key fingerprint verification, so this change
isn't really a regression. Here's an upstream issue:

    https://github.com/puppetlabs/puppetlabs-apt/issues/1175

Before this change, the key fingerprint was only used (by `apt::key`) to
name the file where the key was stored. Because no verification is made
either here or on the `puppetlabs-apt` side, we also remove the
`$gpg_key_id` class param. This change avoids the need to maintain the
hardcoded fingerprint, while also avoiding a false impression that they
fingerprint is somehow verified.

Discussion: https://github.com/voxpupuli/puppet-jenkins/pull/1130